+++ stride-application/src/main/java/io/descoped/stride/application/Application.java
@@ -10,8 +10,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.descoped.stride.application.config.ApplicationConfiguration;
+import io.descoped.stride.application.core.DefaultExceptionServletFilter;
+import io.descoped.stride.application.core.InstanceFactory;
 import io.descoped.stride.application.cors.ApplicationCORSServletFilter;
-import io.descoped.stride.application.factory.InstanceFactory;
 import io.descoped.stride.application.openapi.ApplicationOpenApiResource;
 import io.descoped.stride.application.openapi.ApplicationOpenApiSpecFilter;
 import io.dropwizard.metrics.jetty11.InstrumentedConnectionFactory;
