+++ stride-application/src/main/java/io/descoped/stride/application/core/DefaultConfigurationPostPopulatorProcessor.java
@@ -1,4 +1,4 @@
-package io.descoped.stride.application;
+package io.descoped.stride.application.core;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
