+++ stride-application/src/test/java/io/descoped/stride/application/EmbeddedApplicationTest.java
@@ -5,7 +5,8 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.descoped.stride.application.config.ApplicationConfiguration;
-import io.descoped.stride.application.factory.InstanceFactory;
+import io.descoped.stride.application.core.InstanceFactory;
+import io.descoped.stride.application.core.Logging;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
